refactor : AccommodationRequest , AccommodationController 정리
